Some suggestion for about page
In the "technical details" section, there is word description about "he right to legal prosecution". In the laws of Taiwan (R.O.C), there is no such thing as "the right to legal prosecution". So, I proposed to modify the word like "will file a lawsuit" or "will take legal actions against".